Base game - done
- Приступаем к реализации конкретных штук

+++ Assets/Scripts/General/Wave And Balance System/Balance/PlayerBalanceObserver.cs
@@ -94,7 +94,7 @@ namespace Sampo.Balance
 
         private void RecalcParts() 
         {
-            int sum = buildingsInfl + playerCharacterInfl + sampoUnitsInfl;
+            float sum = buildingsInfl + playerCharacterInfl + sampoUnitsInfl;
             buildingsPart = buildingsInfl / sum;
             playerCharacterPart = playerCharacterInfl / sum;
             sampoUnitsPart = sampoUnitsInfl / sum;
